+++ src/TowerDefense/modele/tourelle/Tourelle.java
@@ -11,7 +11,7 @@ public class Tourelle extends Tours {
 	
 	
 
-	public Tourelle(double x, double y, Terrain terrain, Jeu jeu) {
+	public Tourelle(int x, int y, Terrain terrain, Jeu jeu) {
 		super(x, y, terrain,jeu);
 	}
