+++ src/TowerDefense/modele/tourelle/TourelleBasique.java
@@ -5,7 +5,7 @@ import TowerDefense.modele.Terrain;
 
 public class TourelleBasique extends Tourelle {
 
-	public TourelleBasique(double x, double y, Terrain terrain, Jeu jeu) {
+	public TourelleBasique(int x, int y, Terrain terrain, Jeu jeu) {
 		super(x, y, terrain, jeu);
 		
 	}
